requirements.txt
allure-pytest==2.8.12
pytest==5.3.2
selenium==3.141.0
Faker==4.0.0
mysql-connector-python==8.0.19